backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-multipart>=0.0.20
beautifulsoup4>=4.12.3
python-dotenv>=1.0.1
aiosqlite>=0.20.0
requests>=2.31.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
pandas>=2.0.0
lxml>=4.9.0
